requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.14
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiokafka==0.13.0
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   google-genai
    #   httpx
async-timeout==5.0.1
    # via aiokafka
c2pa-python==0.28.0
    # via -r requirements.in
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via
    #   cryptography
    #   pygit2
charset-normalizer==3.4.4
    # via requests
colorama==0.4.6
    # via pytest
cryptography==46.0.5
    # via
    #   c2pa-python
    #   google-auth
distro==1.9.0
    # via google-genai
google-auth[requests]==2.49.0.dev0
    # via google-genai
google-genai==1.66.0
    # via -r requirements.in
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via google-genai
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
iniconfig==2.3.0
    # via pytest
liboqs-python==0.14.1
    # via -r requirements.in
packaging==26.0
    # via
    #   aiokafka
    #   pytest
    #   wheel
pluggy==1.6.0
    # via pytest
pyasn1==0.6.2
    # via pyasn1-modules
pyasn1-modules==0.4.2
    # via google-auth
pycparser==3.0
    # via cffi
pydantic==2.12.5
    # via
    #   -r requirements.in
    #   google-genai
pydantic-core==2.41.5
    # via pydantic
pygit2==1.19.1
    # via -r requirements.in
pygments==2.19.2
    # via pytest
pytest==9.0.2
    # via c2pa-python
pyyaml==6.0.3
    # via -r requirements.in
requests==2.32.5
    # via
    #   c2pa-python
    #   google-auth
    #   google-genai
sniffio==1.3.1
    # via google-genai
tenacity==9.1.4
    # via google-genai
toml==0.10.2
    # via c2pa-python
typing-extensions==4.15.0
    # via
    #   aiokafka
    #   google-genai
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
urllib3==2.6.3
    # via requests
websockets==16.0
    # via google-genai
wheel==0.46.3
    # via c2pa-python

# The following packages are considered to be unsafe in a requirements file:
# setuptools
